refactor(quality): split modules to clear quality-gate warnings
